# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "toTrack",
    "version": "6.2",
    "description": "toTrack GPS tracking server API documentation. To use the API you need to have a server instance.",
    "contact": {
      "name": "toTrack Support",
      "url": "https://#",
      "email": "support@totrack"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "https://demo.traccar.org/api",
      "description": "Demo Server 1"
    },
    {
      "url": "https://demo2.traccar.org/api",
      "description": "Demo Server 2"
    },
    {
      "url": "https://demo3.traccar.org/api",
      "description": "Demo Server 3"
    },
    {
      "url": "https://demo4.traccar.org/api",
      "description": "Demo Server 4"
    },
    {
      "url": "https://server.traccar.org/api",
      "description": "Subscription Server"
    },
    {
      "url": "http://{host}:{port}/api",
      "description": "Other Server",
      "variables": {
        "host": {
          "default": "localhost"
        },
        "port": {
          "enum": [
            "8082",
            "80"
          ],
          "default": "8082"
        }
      }
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
...
  "paths": {
...
    "/session": {
      "get": {
        "summary": "Fetch Session information",
        "tags": [
          "Session"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {}
          }
        }
      },
      "post": {
        "summary": "Create a new Session",
        "tags": [
          "Session"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        }
      },
      "delete": {
        "summary": "Close the Session",
        "tags": [
          "Session"
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        }
      }
    },
    "/session/openid/auth": {
      "get": {
        "summary": "Fetch Session information",
        "tags": [
          "Session"
        ],
        "parameters": [
          {
          }
        ],
        "responses": {
          "303": {
            "description": "Redirect to OpenID Connect identity provider",
            "content": { }
          }
        }
      }
    },
    "/session/openid/callback": {
      "get": {
        "summary": "OpenID Callback",
        "tags": [
          "Session"
        ],
        "parameters": [
          {
          }
        ],
        "responses": {
          "303": {
            "description": "Successful authentication, redirect to homepage",
            "content": { }
          }
        }
      }
    },
...
    "/notifications": {
      "get": {
        "summary": "Fetch a list of Notifications",
        "description": "Without params, it returns a list of Notifications the user has access to",
        "tags": [
          "Notifications"
        ],
        "parameters": [
          {
            "name": "all",
            "in": "query",
            "description": "Can only be used by admins or managers to fetch all entities",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "Standard users can use this only with their own _userId_",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "deviceId",
            "in": "query",
            "description": "Standard users can use this only with _deviceId_s, they have access to",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "groupId",
            "in": "query",
            "description": "Standard users can use this only with _groupId_s, they have access to",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a Notification",
        "tags": [
          "Notifications"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Notification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
    "/notifications/test": {
      "post": {
        "summary": "Send test notification to current user via Email and SMS",
        "tags": [
          "Notifications"
        ],
        "responses": {
          "204": {
            "description": "Successful sending",
            "content": {}
          },
          "400": {
            "description": "Could happen if sending has failed",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "readonly": {
            "type": "boolean"
          },
          "administrator": {
            "type": "boolean"
          },
          "map": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "zoom": {
            "type": "integer"
          },
          "password": {
            "type": "string"
          },
          "twelveHourFormat": {
            "type": "boolean"
          },
          "coordinateFormat": {
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          },
          "expirationTime": {
            "type": "string",
            "description": "in IS0 8601 format. eg. `1963-11-22T18:30:00Z`",
            "format": "date-time"
          },
          "deviceLimit": {
            "type": "integer"
          },
          "userLimit": {
            "type": "integer"
          },
          "deviceReadonly": {
            "type": "boolean"
          },
          "limitCommands": {
            "type": "boolean"
          },
          "fixedEmail": {
            "type": "boolean"
          },
          "poiLayer": {
            "type": "string"
          },
          "attributes": {
            "type": "object",
            "properties": {}
          }
        }
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "type": "string"
          },
          "always": {
            "type": "boolean"
          },
          "web": {
            "type": "boolean"
          },
          "mail": {
            "type": "boolean"
          },
          "sms": {
            "type": "boolean"
          },
          "calendarId": {
            "type": "integer"
          },
          "attributes": {
            "type": "object",
            "properties": {}
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "description": "Basic HTTP authorization with _email_ and _password_",
        "scheme": "basic"
      }
    }
  }
}